fix: Fix bugs emerging from incomplete merge to dev from feature/water_temperature branch

+++ lib/drivers/water_temperature.h
@@ -8,4 +8,4 @@
 
 void water_temperature_init();
 
-double water_temperature_get();
+int water_temperature_get();
